chore(ast): use `u32` for integer sizes

+++ petal-ast/src/type.rs
@@ -60,7 +60,7 @@ impl Display for TypeKind {
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ResolvedTypeKind {
     /// An integer type.
-    Integer(usize),
+    Integer(u32),
 
     /// The 'void' type.
     Void,
